Change size of AnimationPlayer::AnimationPlayId so an std::optional of it fits in 8 bytes (hardware debug support)

+++ src/engine/core/include/halley/graphics/sprite/animation_player.h
@@ -13,7 +13,7 @@ namespace Halley
 	class AnimationPlayer
 	{
 	public:
-		using AnimationPlayId = size_t;
+		using AnimationPlayId = uint32_t;
 
 		explicit AnimationPlayer(std::shared_ptr<const Animation> animation = std::shared_ptr<const Animation>(), const String& sequence = "default", const String& direction = "default");
